Further chips.js cleanup
