+++ setup/db/DatabaseLoader.js
@@ -29,12 +29,14 @@ const loadsheet = function(sheet) {
 			// load sheet data into some nice data structure and put it in `ex`
 		}
 		else if (sheet.title == 'members') {
+      console.log("Now loading sinx members");
       rows.forEach(r => {
+        console.log("Got a row");
         if(ex.sinxUsers.get(r.userid)==null)
           ex.sinxUsers.set(r.userid,r);
         else
           ex.sinxUsers.get(r.userid).points=parseInt(ex.sinxUsers.get(r.userid).points,10)+parseInt(r.points,10);
-
+          console.log("new pts action");
         console.log(`Loaded user pts action for user: <@${r.userid}>, ${r.points} points, approved by: <@${r.managerid}>`);
       });
 		}
